refactor: Split TypeFactory.swift into discrete files by factory
